Move Bundle ext to CommonUtils
